Check to see that goals are reachable by robot
This commit, particularly,`test_valid_goal` in test_benchmarks.py
checks to see whether the goal state provided to the Sawyer robot
is reachable by the Sawyer's end effector.

It is a WIP, because the IK solver that is used cannot work around
objects in an environment, such as a drawer.