Scripted policies utils and Hammer example
This commit introduces the scripted policy, a tool
that can be used for determining whether a task is
achieveable using the same actions given scripted states.
it will allow contributors to make changes to environments
or dependencies and see if their changes cause any serious breaks
but without having to run a suite of single task rl benchmarks
on the environments that they have made changes to.

In this commit is scripted policies utils, as well as an example
of how to use them on the environment hammer-v1. The same relative
design pattern should be followed, along with the same parameters
to the environments (random_init=True)

Lastly, a parameter called render should only be set to true for
debugging purposes, but in the test suite it should be set to false.